% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsAccession.R
\name{ngsAccession}
\alias{ngsAccession}
\title{Accession NGS sequence metadata to SQL Server table}
\usage{
ngsAccession(df)
}
\arguments{
\item{df}{formatted indexing data frame resulting 
from \code{\link{ngsFormatDF}}.}
}
\description{
Accession NGS sequence metadata to SQL Server table.
}
\seealso{
\code{\link{ngsFormatDF}}, \code{\link{ngsRename}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
