% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsFormatDF.R
\name{ngsFormatDF}
\alias{ngsFormatDF}
\title{Format NGS data frame}
\usage{
ngsFormatDF(
  library.name,
  df,
  folder,
  species = "species",
  f.index = "F_index",
  r.index = "R_index",
  labid = "LabID",
  filename = "filename",
  read.direction = "Read_Direction"
)
}
\arguments{
\item{library.name}{name of run library}

\item{df}{original library data frame}

\item{folder}{folder where original FASTQ files are stored}

\item{species}{species name}

\item{f.index}{forward index}

\item{r.index}{reverse index}

\item{labid}{SWFSC LABID}

\item{filename}{filename of original FASTQ file}

\item{read.direction}{direction of reads}
}
\description{
Format NGS data frame for accessioning. Assigns correct column 
  names, checks that files found.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
