% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsRename.R
\name{ngsRename}
\alias{ngsRename}
\title{Rename NGS files}
\usage{
ngsRename(
  df,
  old.folder,
  new.folder,
  leave.files = TRUE,
  use.archive.folders = TRUE
)
}
\arguments{
\item{df}{formatted indexing data frame resulting 
from \code{\link{ngsFormatDF}}.}

\item{old.folder}{folder where original files reside.}

\item{new.folder}{folder where renamed files should be placed.}

\item{leave.files}{leave files in \code{old.folder}? If \code{TRUE} a copy 
operation will take place, leaving files in {old.folder}. If \code{FALSE},
the a rename operation will take place. If \code{old.folder} is 
different from \code{new.folder} or \code{use.archive.folders} is 
\code{FALSE} this will be the same as moving the files.}

\item{use.archive.folders}{If \code{TRUE} the target folder will be 
'\code{<new.folder>}/Gspp/run.library' where 'Gspp' is the genus and 
species abbreviation. If \code{FALSE}, the target folder will just be 
the root of \code{new.folder}.}
}
\description{
Rename NGS files using new accessioned names from indexing
  data frame.
}
\seealso{
\code{\link{ngsFormatDF}}, \code{\link{ngsAccession}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
