% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileupCall.R
\name{pileupCallRun}
\alias{pileupCallRun}
\alias{pileupCallFile}
\title{Call bases from pileup file}
\usage{
pileupCallRun(
  min.cov.call,
  min.cov.freq,
  min.base.freq,
  min.ins.freq,
  min.prob.freq,
  min.binom.prob,
  folder = ".",
  pattern = "\\\\.pileup$",
  label = NULL,
  num.cores = NULL
)

pileupCallFile(
  fname,
  min.cov.call,
  min.cov.freq,
  min.base.freq,
  min.ins.freq,
  num.cores = NULL
)
}
\arguments{
\item{min.cov.call}{minimum coverage for base calling. Sites with coverage below
this are assigned \code{N}'s.}

\item{min.cov.freq}{minimum coverage above which \code{min.freq} is applied. 
Sites below this value and >= than \code{min.cov} will only be called 
if all reads agree.}

\item{min.base.freq}{minimum frequency of either the reference or alternate base
for calling. If both bases are below this frequency, an \code{N} 
is assigned.}

\item{min.ins.freq}{minimum frequency of insertion.}

\item{min.prob.freq}{minimum frequency for binomial probability.}

\item{min.binom.prob}{minimum probability from binomial distribution.}

\item{folder}{folder containing pileup files from a run}

\item{pattern}{text pattern for pileup files.}

\item{label}{label for run output files.}

\item{num.cores}{number of cores to use during processing. If \code{NULL}, 
will default to \code{parallel::detectCores() - 1}.}

\item{fname}{filename of pileup file}
}
\value{
list with the following elements:
\tabular{ll}{ 
  \code{plp} \tab data frame.\cr 
  \code{base.lo} \tab data frame of bases and log-odds.\cr 
}
}
\description{
Reads a pileup formatted file created by \code{samtools mpileup}
  and calls bases for each chromosome listed. Base calling is controlled
  by coverage and frequency parameters
}
\note{
The input pileup file should be the result of a call to 
  \code{samtools mpileup} on a single BAM file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
