% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileupConsensus.R
\name{pileupConsensus}
\alias{pileupConsensus}
\title{Create consensus sequences from pileup file}
\usage{
pileupConsensus(fname, min.cov, min.freq, min.freq.cov)
}
\arguments{
\item{fname}{filename of pileup file}

\item{min.cov}{minimum coverage for base calling (sites with coverage below
this are assigned N's).}

\item{min.freq}{minimum frequency of either the reference or alternate base
for calling. If both bases are below this frequency, an N is assigned.}

\item{min.freq.cov}{minimum coverage above which min.freq is applied. Sites
below this and >= than min.cov will only be called if all reads agree.}
}
\value{
list with the following elements:
\tabular{ll}{ 
  \code{cons.seq} \tab named list of consensus sequences for each chromosome.\cr 
  \code{plp} \tab data frame of the pileup data with base frequencies.\cr 
  \code{insertions} \tab data frame of insertion locations and frequencies.
    If there are no insertions, this will be \code{NULL}.\cr 
}
}
\description{
Reads a pileup formatted file created by \code{samtools mpileup}
  and creates a consensus sequence for each chromosome listed.
}
\note{
The input pileup file should be the result of a call to 
  \code{samtools mpileup} on a single BAM file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
