% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnSex.R
\name{pinnSex}
\alias{pinnSex}
\title{Determine sex based on Tm thresholds}
\usage{
pinnSex(x, tm.x = c(82, 85), tm.y = c(87, 90))
}
\arguments{
\item{x}{either a filename of an Excel (.xlsx) file or data frame created 
from the file}

\item{tm.x, tm.y}{two-element vectors giving the thresholds for Tm values for
x and y chromosomes.}
}
\description{
Determines sex based on Tm thresholds.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
