% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPileup.R
\name{readPileup}
\alias{readPileup}
\title{Read pileup file}
\usage{
readPileup(fname, all.columns = FALSE)
}
\arguments{
\item{fname}{filename of pileup file}

\item{all.columns}{If \code{FALSE} (default), only the first 6 columns are 
returned. If \code{TRUE}, all columns are returned, but only the six are 
informatively named.}
}
\value{
data frame representing the pileup file with the following columns:
\tabular{ll}{ 
  \code{chrom} \tab Chromosome name.\cr 
  \code{pos} \tab 1-based position on the chromosome.\cr 
  \code{ref} \tab Reference base at this position.\cr 
  \code{cov} \tab Number of reads covering this position.\cr 
  \code{bases} \tab Read bases.\cr 
  \code{quals} \tab Base qualities, encoded as ASCII characters.\cr 
}
}
\description{
Reads a pileup format file created by \code{samtools mpileup}.
}
\note{
The input pileup file should be the result of a call to 
  'samtools mpileup' on a single BAM file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
