% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swfscGenetics-package.R
\docType{package}
\name{swfscGenetics-package}
\alias{swfscGenetics-package}
\alias{swfscGenetics}
\title{Miscellaneous functions for genetic data manipulation}
\description{
swfscGenetics
}
\keyword{package}
