% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsFormatDF.R
\name{ngsFormatDF}
\alias{ngsFormatDF}
\title{Format NGS data frame}
\usage{
ngsFormatDF(
  library.name,
  library.filename = NULL,
  labid = "LabID",
  d.id = "D_id",
  species = "species",
  i7.index = "i7_index",
  i5.index = "i5_index",
  read.direction = "Read_Direction"
)
}
\arguments{
\item{library.name}{name of run library. This should also be the folder 
name where the files to be archived are located.}

\item{library.filename}{filename of library spreadsheet. If \code{NULL}, 
the .csv file in the \code{library.name} folder}

\item{labid}{SWFSC LABID column name}

\item{d.id}{DNA ID column name}

\item{species}{species name column name}

\item{i7.index}{forward index column name}

\item{i5.index}{reverse index column name}

\item{read.direction}{direction of reads column name}
}
\description{
Format NGS data frame for accessioning. Assigns correct column 
  names, checks that files found.
}
\seealso{
\code{\link{ngsRename}}, \code{\link{ngsAccession}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
