% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsRename.R
\name{ngsRename}
\alias{ngsRename}
\title{Rename NGS files}
\usage{
ngsRename(df, old.folder, new.folder, leave.files = TRUE)
}
\arguments{
\item{df}{formatted indexing data frame resulting 
from \code{\link{ngsFormatDF}}.}

\item{old.folder}{folder where original files reside.}

\item{new.folder}{folder where renamed files should be placed.}

\item{leave.files}{leave files in \code{old.folder}? If \code{FALSE} files 
will be moved.}
}
\description{
Rename NGS files using new accessioned names from indexing
  data frame.
}
\seealso{
\code{\link{ngsFormatDF}}, \code{\link{ngsAccession}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
