% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngsRename.R
\name{ngsRename}
\alias{ngsRename}
\title{Rename NGS files}
\usage{
ngsRename(df, folder)
}
\arguments{
\item{df}{formatted indexing data frame resulting 
from \code{\link{ngsFormatDF}}.}

\item{folder}{folder where original files reside.}
}
\description{
Rename NGS files using new accessioned names from indexing
  data frame.
}
\seealso{
\code{\link{ngsFormatDF}}, \code{\link{ngsAccession}}
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
