% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileupConsensus.R
\name{pileupConsensus}
\alias{pileupConsensus}
\title{Create consensus sequences from pileup file}
\usage{
pileupConsensus(fname, min.cov, min.freq, min.freq.cov, num.cores = NULL)
}
\arguments{
\item{fname}{filename of pileup file}

\item{min.cov}{minimum coverage for base calling. Sites with coverage below
this are assigned \code{N}'s.}

\item{min.freq}{minimum frequency of either the reference or alternate base
for calling. If both bases are below this frequency, an \code{N} 
is assigned.}

\item{min.freq.cov}{minimum coverage above which \code{min.freq} is applied. 
Sites below this value and >= than \code{min.cov} will only be called 
if all reads agree.}

\item{num.cores}{number of cores to use during processing. If \code{NULL}, 
will default to \code{parallel::detectCores() - 1}.}
}
\value{
list with the following elements:
\tabular{ll}{ 
  \code{cons.seq} \tab named list of consensus sequences for each chromosome.\cr 
  \code{plp} \tab data frame of the pileup data with base frequencies.\cr 
  \code{insertions} \tab data frame of insertion locations and frequencies.
    If there are no insertions, this will be \code{NULL}.\cr 
}
}
\description{
Reads a pileup formatted file created by \code{samtools mpileup}
  and creates a consensus sequence for each chromosome listed.
}
\note{
The input pileup file should be the result of a call to 
  \code{samtools mpileup} on a single BAM file.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
